@@ -1,11 +1,13 @@
 {
   "name": "lifemonitor",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "scripts": {
-    "start": "node server.js",
-    "dev:start": "ng serve",
-    "dev:build": "ng build --prod",
-    "dev:test": "ng test",
+    "start": "ng build --configuration production && http-server -p 4202 --ssl -C ./certs/lm.crt -K ./certs/lm.key --host lm dist/lifemonitor",
+    "start:dev": "ng serve --port 4200 --ssl --ssl-key ./certs/lm.key --ssl-cert ./certs/lm.crt --host lm",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "test:ci": "ng test --browsers=ChromeHeadless --watch=false",
     "dev:lint": "ng lint",
     "dev:e2e": "ng e2e",
     "postinstall": "ngcc"
@@ -23,45 +25,66 @@
     }
   },
   "dependencies": {
-    "@angular/animations": "11.2.5",
-    "@angular/common": "11.2.5",
-    "@angular/compiler": "11.2.5",
-    "@angular/core": "11.2.5",
-    "@angular/forms": "11.2.5",
-    "@angular/localize": "11.2.5",
-    "@angular/platform-browser": "11.2.5",
-    "@angular/platform-browser-dynamic": "11.2.5",
-    "@angular/router": "11.2.5",
+    "@angular/animations": "12.2.17",
+    "@angular/cdk": "^12.2.13",
+    "@angular/common": "12.2.17",
+    "@angular/compiler": "12.2.17",
+    "@angular/core": "12.2.17",
+    "@angular/forms": "12.2.17",
+    "@angular/localize": "12.2.17",
+    "@angular/platform-browser": "12.2.17",
+    "@angular/platform-browser-dynamic": "12.2.17",
+    "@angular/pwa": "^12.2.18",
+    "@angular/router": "12.2.17",
+    "@angular/service-worker": "^12.2.17",
     "@bity/oauth2-auth-code-pkce": "^2.13.0",
-    "@ng-bootstrap/ng-bootstrap": "9.0.2",
+    "@ng-bootstrap/ng-bootstrap": "~10.0.0",
+    "@types/async-lock": "^1.4.0",
+    "@types/crypto-js": "^4.1.1",
+    "angular-datatables": "^12.0.1",
     "ansi-html": "~0.0.7",
+    "async-lock": "^1.4.0",
     "bootstrap": "4.6.0",
     "bootstrap-select": "^1.13.18",
     "bs-stepper": "~1.7.0",
     "chalk": "~4.1.1",
     "chart.js": "^2.9.4",
     "core-js": "3.9.1",
+    "crypto": "^1.0.1",
+    "crypto-js": "^4.1.1",
     "datatables": "^1.10.18",
+    "datatables.net": "^1.13.4",
+    "datatables.net-bs4": "^1.13.4",
     "express": "4.17.1",
     "hover.css": "^2.3.2",
-    "ng2-charts": "^2.4.2",
+    "idb": "^7.1.1",
+    "jquery": "^3.6.0",
+    "ng2-charts": "^2.4.1",
+    "ngx-cookie-service": "^12.1.0-rc.5",
+    "ngx-socket-io": "^4.1.0",
     "ngx-toastr": "13.2.0",
+    "primeicons": "^6.0.1",
+    "primeng": "^12.2.2",
     "rxjs": "6.6.6",
+    "socket.io": "^4.6.0",
     "strip-ansi": "~7.0.0",
+    "swiped-events": "^1.1.7",
     "tslib": "^2.0.3",
-    "typescript-logger": "^5.0.1",
-    "zone.js": "0.11.4"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.14",
-    "@angular/cli": "^11.2.4",
-    "@angular/compiler-cli": "11.2.5",
-    "@angular/language-service": "11.2.5",
+    "@angular-devkit/build-angular": "^12.2.18",
+    "@angular/cli": "^12.2.18",
+    "@angular/compiler-cli": "12.2.17",
+    "@angular/language-service": "12.2.17",
+    "@types/datatables.net": "^1.10.21",
     "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "2.0.8",
+    "@types/jquery": "^3.5.9",
     "@types/node": "14.14.33",
-    "admin-lte": "3.1.0",
+    "admin-lte": "3.2.0",
     "codelyzer": "^6.0.1",
+    "eslint": "^8.29.0",
     "husky": "5.1.3",
     "jasmine-core": "~3.9.0",
     "jasmine-spec-reporter": "~6.0.0",
@@ -75,10 +98,7 @@
     "pretty-quick": "3.1.0",
     "protractor": "^7.0.0",
     "ts-node": "9.1.1",
-    "tslint": "6.1.3",
-    "tslint-config-prettier": "1.18.0",
-    "tslint-plugin-prettier": "2.3.0",
-    "typescript": "4.0.6",
+    "typescript": "4.3.5",
     "webdriver-manager": "^12.1.8"
   }
 }
