@@ -29,7 +29,8 @@ export class SearchBarComponent implements OnInit {
   @Input() elementType: string;
   @Input() filterValue: string;
   @Input() sortingOrder: string;
-  @Output() browseEnabled = new EventEmitter<boolean>();
+  @Input() browseEnabled: boolean = false;
+  @Output() browseEnabledChange = new EventEmitter<boolean>();
   @Output() filterValueChange = new EventEmitter<string>();
   @Output() sortingOrderChange = new EventEmitter<string>();
   @ViewChild('searchInputText') searchInputText: ElementRef;
@@ -63,16 +64,16 @@ export class SearchBarComponent implements OnInit {
     const title: string = !this.actualFilterValue
       ? 'Browse registered ' + this.elementType
       : 'Search ' + this.elementType + ' by UUID or name';
-    this.searchButton.nativeElement.setAttribute('data-original-title', title);
+    // this.searchButton.nativeElement.setAttribute('data-original-title', title);
   }
 
   public onBrowseButtonClick($event) {
     this.browseButtonEnabled = !this.browseButtonEnabled;
     this.logger.debug('OnShowAllChange', $event, this.browseButtonEnabled);
-    this.browseEnabled.emit(this.browseButtonEnabled);
+    this.browseEnabledChange.emit(this.browseButtonEnabled);
     if (this.browseButtonEnabled) {
       this._searchAllValues = true;
-      this.emitValue(this.actualFilterValue);
+      // this.emitValue(this.actualFilterValue);
     } else {
       this.reset();
     }
