@@ -1,14 +1,23 @@
 import {
+  AfterViewChecked,
+  ChangeDetectionStrategy,
   Component,
   EventEmitter,
   Input,
+  NgZone,
   OnChanges,
   OnInit,
   Output,
   SimpleChanges,
 } from '@angular/core';
 import { Router } from '@angular/router';
-import { ChartData, ChartDataSets, ChartOptions, ChartType } from 'chart.js';
+import {
+  Chart,
+  ChartData,
+  ChartDataSets,
+  ChartOptions,
+  ChartType,
+} from 'chart.js';
 import { Label } from 'ng2-charts';
 import { DateUtils } from 'src/app/models/common.models';
 import { StatusStatsItem } from 'src/app/models/stats.model';
@@ -19,8 +28,10 @@ import { Logger, LoggerManager } from 'src/app/utils/logging';
   selector: 'stats-bar-chart',
   templateUrl: './stats-bar-chart.component.html',
   styleUrls: ['./stats-bar-chart.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class StatsBarChartComponent implements OnInit, OnChanges {
+export class StatsBarChartComponent
+  implements OnInit, OnChanges, AfterViewChecked {
   @Input() stats!: StatusStatsItem[];
   @Output() selectedItem = new EventEmitter<StatusStatsItem>();
 
@@ -34,10 +45,9 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
       xAxes: [{ display: false }],
       yAxes: [
         {
-          ticks: {
-            minor: {
-            }
-          },
+          // ticks: {
+          //   minor: {},
+          // },
           display: false,
           time: {
             unit: 'second',
@@ -57,9 +67,9 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
         // Create element on first render
         if (!tooltipEl) {
           tooltipEl = document.createElement('div');
-          tooltipEl.style.background = "black";
-          tooltipEl.style.color = "white";
-          tooltipEl.style.fontWeight = "200";
+          tooltipEl.style.background = 'black';
+          tooltipEl.style.color = 'white';
+          tooltipEl.style.fontWeight = '200';
           tooltipEl.id = 'chartjs-tooltip';
           tooltipEl.style.borderRadius = '10px';
           tooltipEl.innerHTML = '<table></table>';
@@ -68,7 +78,7 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
 
         // Hide if no tooltip
         if (tooltipModel.opacity === 0) {
-          tooltipEl.style.opacity = "0";
+          tooltipEl.style.opacity = '0';
           return;
         }
 
@@ -80,7 +90,7 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
           tooltipEl.classList.add('no-transform');
         }
 
-        function getBody(bodyItem: { lines: any; }) {
+        function getBody(bodyItem: { lines: any }) {
           return bodyItem.lines;
         }
 
@@ -105,7 +115,8 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
           });
           innerHtml += '</tbody>';
           // append footer
-          var footer = '<div style="color: lightgray; font-size: 8pt; text-align: right;">';
+          var footer =
+            '<div style="color: lightgray; font-size: 8pt; text-align: right;">';
           footer += tooltipModel.footer;
           footer += '</div>';
           innerHtml += footer;
@@ -118,25 +129,35 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
         var position = this._chart.canvas.getBoundingClientRect();
 
         // Display, position, and set styles for font
-        tooltipEl.style.opacity = "1";
+        tooltipEl.style.opacity = '1';
         tooltipEl.style.position = 'absolute';
-        tooltipEl.style.left = position.left + window.pageXOffset + tooltipModel.caretX + 'px';
-        tooltipEl.style.top = position.top + window.pageYOffset + tooltipModel.caretY + 'px';
+        tooltipEl.style.left =
+          position.left + window.pageXOffset + tooltipModel.caretX + 'px';
+        tooltipEl.style.top =
+          position.top + window.pageYOffset + tooltipModel.caretY + 'px';
         tooltipEl.style.fontFamily = tooltipModel._bodyFontFamily;
         tooltipEl.style.fontSize = tooltipModel.bodyFontSize + 'px';
         tooltipEl.style.fontStyle = tooltipModel._bodyFontStyle;
-        tooltipEl.style.padding = tooltipModel.yPadding + 'px ' + tooltipModel.xPadding + 'px';
+        tooltipEl.style.padding =
+          tooltipModel.yPadding + 'px ' + tooltipModel.xPadding + 'px';
         tooltipEl.style.pointerEvents = 'none';
       },
       callbacks: {
         title: (tooltipItem, data: ChartData) => {
           let item: TestBuild = this.getItemTestBuild(tooltipItem[0], data);
-          return '<span style="color: transparent; font-size: 0.7rem; text-shadow: 0 0 0 ' + item.color + '; ">&#9899;</span>' +
-            ' Build #' + item.build_id + ' : ' + item.status;
+          return (
+            '<span style="color: transparent; font-size: 0.7rem; text-shadow: 0 0 0 ' +
+            item.color +
+            '; ">&#9899;</span>' +
+            ' Build #' +
+            item.build_id +
+            ' : ' +
+            item.status
+          );
         },
         label: (tooltipItem, data: ChartData) => {
           let item: TestBuild = this.getItemTestBuild(tooltipItem, data);
-          this.logger.debug("Check stats item", tooltipItem, data);
+          this.logger.debug('Check stats item', tooltipItem, data);
           let label = data.labels[tooltipItem.index];
           let sec = data.datasets[tooltipItem.datasetIndex].data[
             tooltipItem.index
@@ -148,17 +169,34 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
           if (hours > 0) duration += hours + 'h ';
           if (minutes > 0) duration += minutes + 'm ';
           if (sec >= 0) duration += seconds + 's';
-          this.logger.debug("DURATION", hours, minutes, seconds, duration, tooltipItem, item, data);
-          let ltext = '<div class="ml-1"><i class="far fa-calendar-check"></i> started: '
-            + DateUtils.formatTimestamp(item.timestamp.toString()) + '</div>';
-          ltext += '<div class="ml-1"><i class="fas fa-stopwatch"></i> duration: '
-            + duration + '</div>'
+          this.logger.debug(
+            'DURATION',
+            hours,
+            minutes,
+            seconds,
+            duration,
+            tooltipItem,
+            item,
+            data
+          );
+          let ltext =
+            '<div class="ml-1"><i class="far fa-calendar-check"></i> started: ' +
+            DateUtils.formatTimestamp(item.timestamp.toString()) +
+            '</div>';
+          ltext +=
+            '<div class="ml-1"><i class="fas fa-stopwatch"></i> duration: ' +
+            duration +
+            '</div>';
           return ltext;
         },
         footer: (tooltipItem, data) => {
           let index = tooltipItem[0]['datasetIndex'];
           let item = this.stats[index] as TestBuild;
-          return 'click to see on <b class="text-white">' + item.instance.service.type + '</b>';
+          return (
+            'click to see on <b class="text-white">' +
+            item.instance.service.type +
+            '</b>'
+          );
         },
       },
     },
@@ -190,14 +228,48 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
 
   public getItemTestBuild(tooltipItem: any, data: ChartData): TestBuild {
     if (!tooltipItem || !data) return null;
-    this.logger.debug("item, data", tooltipItem, data);
+    this.logger.debug('item, data', tooltipItem, data);
     let index = tooltipItem['datasetIndex'];
     return this.stats[index] as TestBuild;
   }
 
   public selectedObject: StatusStatsItem;
 
-  constructor(private router: Router) { }
+  constructor(private router: Router, private zone: NgZone) {}
+
+  public barChartData: ChartDataSets[] = [];
+
+  ngOnInit(): void {}
+
+  ngAfterViewChecked(): void {}
+
+  ngOnChanges(changes: SimpleChanges): void {
+    this.zone.runOutsideAngular(() => {
+      this.refresh();
+    });
+  }
+
+  refresh() {
+    // this.logger.debug('Stats ', this.stats);
+    this.barChartData = [];
+    this.barColors = [];
+    this.logger.debug('Current bar stas: ', this.stats);
+    let max: number = Math.max.apply(
+      Math,
+      this.stats.map(function (o) {
+        return o.duration || 0;
+      })
+    );
+    this.logger.debug('MAX duration', max);
+    for (let i in this.stats) {
+      let build: StatusStatsItem = this.stats[i];
+      this.barChartData.push({
+        data: [build.duration || 0],
+        label: 'duration',
+      });
+      this.barColors.push(this.getColor(build.status));
+    }
+  }
 
   private getColor(label: string) {
     return {
@@ -212,28 +284,6 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
     };
   }
 
-  public barChartData: ChartDataSets[] = [];
-
-  ngOnInit(): void { }
-
-  ngOnChanges(changes: SimpleChanges): void {
-    // this.logger.debug('Stats ', this.stats);
-    this.barChartData = [];
-    this.barColors = [];
-    this.logger.debug("Current bar stas: ", this.stats);
-    let max: number = Math.max.apply(Math,
-      this.stats.map(function (o) { return o.duration || 0; }))
-    this.logger.debug("MAX duration", max);
-    for (let i in this.stats) {
-      let build: StatusStatsItem = this.stats[i];
-      this.barChartData.push({
-        data: [build.duration || 0],
-        label: 'duration',
-      });
-      this.barColors.push(this.getColor(build.status));
-    }
-  }
-
   // events
   public chartClicked({
     event,
@@ -242,7 +292,7 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
     event: MouseEvent;
     active: { _chart: any }[];
   }): void {
-    this.logger.debug("Chart click event", event, active);
+    this.logger.debug('Chart click event', event, active);
     if (!active || active.length == 0) return;
     let chart: Chart = active[0]._chart;
     let selectedElements: [{}] = chart.getElementAtEvent(event);
@@ -265,6 +315,6 @@ export class StatsBarChartComponent implements OnInit, OnChanges {
     event: MouseEvent;
     active: {}[];
   }): void {
-    this.logger.debug("chartHovered Event", event, active);
+    this.logger.debug('chartHovered Event', event, active);
   }
 }
